+++ server/src/test/java/com/epam/aidial/core/server/function/BaseFunctionTest.java
@@ -1,5 +1,6 @@
 package com.epam.aidial.core.server.function;
 
+import com.epam.aidial.core.server.util.ChatUtil;
 import com.epam.aidial.core.server.util.ProxyUtil;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.junit.jupiter.api.Test;
@@ -129,7 +130,7 @@ public class BaseFunctionTest {
                 }
                 """;
         ObjectNode tree = (ObjectNode) ProxyUtil.MAPPER.readTree(content.getBytes());
-        Set<String> result = ProxyUtil.collectAttachments(tree, List.of("@.messages[*].custom_content.attachments[*].url",
+        Set<String> result = ChatUtil.collectAttachments(tree, List.of("@.messages[*].custom_content.attachments[*].url",
                 "@.messages[*].custom_content.stages[*].attachments[*].url",
                 "@.messages[*].content[*].image_url.url",
                 "@.url",
